requirements.txt
annotated-types==0.7.0
anyio~=4.6.2.post1
click~=8.1.7
exceptiongroup==1.2.2
fastapi~=0.115.4
h11==0.14.0
idna==3.10
numpy~=2.0.2
opencv-python~=4.10.0.84
Pillow~=11.0.0
pydantic~=2.9.2
pydantic_core~=2.23.4
python-multipart~=0.0.12
sniffio==1.3.1
starlette~=0.41.2
typing_extensions==4.12.2
uvicorn~=0.32.0

defusedxml~=0.7.1
cffi~=1.17.1
future~=1.0.0
ipython~=8.12.3
pip~=23.2.1
attrs~=24.2.0
argparse~=1.4.0
cryptography~=43.0.3
filelock~=3.16.1
redis~=5.2.0
setuptools~=75.2.0
pytz~=2024.2
numba~=0.60.0
requests~=2.32.3
websockets~=12.0
orjson~=3.10.10
itsdangerous~=2.2.0
PyYAML~=6.0.2
Jinja2~=3.1.4
httpx~=0.27.2
oss2~=2.19.1
pycryptodome
apscheduler~=3.10.4